requirements.txt
# Video and Image Processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Audio Processing and Separation
demucs>=4.0.0
pydub>=0.25.1

# AI Transcribing and LLM Communication
openai-whisper>=20231117
requests>=2.31.0

# Production Readiness (Env and Resilience)
python-dotenv>=1.0.0
tenacity>=8.2.0

# PyTorch and Music Generation (AudioCraft)
torch>=2.1.0
torchaudio>=2.1.0
xformers>=0.0.22
git+https://github.com/facebookresearch/audiocraft.git

# Web Application
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Authentication & Database
SQLAlchemy>=2.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Public Website Tunneling
pyngrok>=7.0.0